Remove solver.
