--- torchbenchmark/models/speech_transformer/solver/__init__.py
@@ -1,1 +1,0 @@
-from solver.solver import *
